refactor(dice): extract DiceTestbench script into separate modules
Split the 364-line script tag into four focused files:
- DiceTestbench.types.ts — type exports
- DiceTestbench.constants.ts — templateShapeOptions
- DiceTestbench.utils.ts — categorizeDie, parseFormulaForDice
- DiceTestbench.state.svelte.ts — reactive state via createDiceTestbenchState()

Also fixes the attackUtils import path and adds missing
primaryDieValue/primaryDieModifier to the stageAndRoll options.
