+++ src/view/debug/DiceTestbench.constants.ts
@@ -1,0 +1,8 @@
+export const templateShapeOptions = [
+	{ value: '', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.none' },
+	{ value: 'circle', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.circle' },
+	{ value: 'cone', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.cone' },
+	{ value: 'emanation', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.emanation' },
+	{ value: 'line', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.line' },
+	{ value: 'square', labelKey: 'NIMBLE.diceTestbench.rollBuilder.templateShape.square' },
+];
